requirements.txt
coreapi==2.3.3
coreschema==0.0.4
django-allauth==0.45.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.12.4
celery==5.1.2